Add support for using pre-built lunr serialized index
Search.html now looks for a prebuilt.json file to load, and
if a 404 is returned continues on the former process of loading
index.json and building the index.

Overall, using a prebuilt index on the example site was approx
2.7x faster than building in the browser, even accounting for the
larger file size.

Also included a shell script which builds hugo to and then prebuilds
lunr index file. Note: This requires node and lunr module installed
locally, and the package.json file is NOT being included in this.