Add support for using a prebuild lunr index with fallback
Search.html now looks for a prebuilt.json file to load, and if a 404
is returned falls back to loading index.json and building an index.

Overall, using a prebuilt index on the example site was approx 2.7x
faster than building in the browser, even accounting for the larger
file size.

To receive both URI and title in search results, modified index.json
to include a joined field used as lunr ref, so pageIndex no longer
needs to be a global page index holding all of the JSON just to retrieve
title on search results. This was required as that data was no longer
available with prebuild index, but overall should be a gain as well.

Updates the version of lunr that is loaded to match the
prebuild index version included from node.js. Lunr gives a warning if
there is a mismatch, but it still works.

Also included a shell script which builds hugo to and then prebuilds
lunr index file. Note: This requires node and lunr module installed
locally, and the package.json file is NOT included.